Move vault helpers into own module
