+++ main.py
@@ -1,18 +1,18 @@
-import base64
 import json
 import os
 import sys
-import urllib.parse
-import tempfile
 import time
 import traceback
-from pathlib import PurePosixPath
 
-import hvac
-import requests
 from google.cloud import storage
 
-from honeyflare import create_libhoney_client, process_bucket_object, RetriableError, logfmt
+from honeyflare import (
+    create_libhoney_client,
+    process_bucket_object,
+    RetriableError,
+    logfmt,
+    vault,
+)
 
 storage_client = storage.Client()
 
@@ -34,61 +34,6 @@ sampling_rate_by_status = {int(key): val for key, val in json.loads(os.environ.g
 lock_bucket = os.environ.get('LOCK_BUCKET')
 
 
-def get_vault_secret(vault_url):
-    '''
-    Fetch honeycomb API token from Vault.
-
-    The url should be on the form
-        vault://<host>:<port>/<mount-point>/<secret-path>?key=<lookup-key>&ca=<urlsafe-b64-ca-cert>[&https=false]
-    '''
-    parsed_url = urllib.parse.urlparse(vault_url)
-    parsed_parameters = urllib.parse.parse_qs(parsed_url.query)
-
-    scheme = 'https'
-    https = parsed_parameters.get('https')
-    if https == 'false':
-        scheme = 'http'
-
-    b64_ca_cert = parsed_parameters.get('ca')
-    if b64_ca_cert:
-        ca_cert = base64.urlsafe_b64decode(b64_ca_cert[0].encode('utf-8'))
-        with tempfile.NamedTemporaryFile(delete=False) as fh:
-            fh.write(ca_cert)
-            os.environ['REQUESTS_CA_BUNDLE'] = fh.name
-
-    vault_url = '%s://%s' % (scheme, parsed_url.netloc)
-    vault_role = parsed_parameters.get('role', 'honeyflare')
-    client = hvac.Client(url=vault_url)
-
-    client.auth.gcp.login(role=vault_role, jwt=get_auth_jwt(vault_role))
-
-    parsed_path = PurePosixPath(parsed_url.path)
-
-    honeycomb_key = client.secrets.kv.read_secret_version(
-        path='/'.join(parsed_path.parts[2:]),
-        mount_point=parsed_path.parts[1],
-    )
-    key = parsed_parameters['key'][0]
-
-    # Clean up to not affect other https queries
-    del os.environ['REQUESTS_CA_BUNDLE']
-
-    return honeycomb_key['data']['data'][key]
-
-
-def get_auth_jwt(vault_role):
-    response = requests.get('http://metadata/computeMetadata/v1/instance/service-accounts/default/identity',
-        params={
-            'audience': 'vault/%s' % vault_role,
-        },
-        headers={
-            'Metadata-Flavor': 'Google',
-        },
-    )
-    response.raise_for_status()
-    return response.text
-
-
 def main(event, context):
     '''
     Triggered by a change to a Cloud Storage bucket.
@@ -107,7 +52,7 @@ def main(event, context):
     start_time = time.time()
     try:
         if honeycomb_key.startswith('vault://'):
-            honeycomb_key = get_vault_secret(honeycomb_key)
+            honeycomb_key = vault.get_vault_secret(honeycomb_key)
 
         bucket = storage_client.bucket(event['bucket'])
         events_handled = process_bucket_object(
